UqsWeather/Uqs.Weather.Tests.Unit/WeatherForecastControllerTests.cs
using Microsoft.Extensions.Logging.Abstractions;
using Uqs.Weather.Controllers;

namespace Uqs.Weather.Tests.Unit;

public class WeatherForecastControllerTests
{
    [Theory]
    [InlineData(-100, -148)]
    [InlineData(-10.1, 13.8)]
    [InlineData(10, 50)]
    public void ConvertCToF_ReceivesDouble_CorrectlyConverts(double c, double f)
    {

        var logger = NullLogger<WeatherForecastController>.Instance;
        var controller = new WeatherForecastController(logger, null!, null!, null!);
        double actual = controller.ConvertCToF(c);
        Assert.Equal(f, actual, 1);
    }

    [Fact]
    public async void GetReal_NotInterestedInTodayWeather_WFStartFromNextDay()
    {
        // Arrange
        const double nextDayTemp = 3.3;
        const double day5Temp = 7.7;
        var today = new DateTime(2022, 1, 1);
        var realWeatherTemps = new double[] { 2, nextDayTemp, 4, 5.5, 6, day5Temp, 8 };
        var clientStub = new ClientStub(today, realWeatherTemps);
        var controller = new WeatherForecastController(null!, clientStub, null!, null!);

        // Act
        IEnumerable<WeatherForecast> wfs = await controller.GetReal();
        // Assert
        Assert.Equal(3, wfs.First().TemperatureC);
    }
}
